+++ CHANGELOG.md
@@ -8,6 +8,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/inspirum/balikobot-php/compare/v7.12.0...master)
 - Added **ZASILKOVNA_BG_BOXNOW_BOX** service type
+- Added **ZASILKOVNA_CY_BOXNOW_BOX** service type
+- Added **ZASILKOVNA_HR_BOXNOW_BOX** service type
+- Added **ZASILKOVNA_RO_FAN_BOX** service type
+- Added **PBH_CP_NB** service type
+- Added **DHLFREIGHTEC_ELD** service type
+- Added **DHLFREIGHTEC_ELI** service type
+- Added **INPOST** carrier
+- Added support for new options from documentation **v2.017** (2025-04-01)
+  - Added **rma_association** attributes
+  - Added **p_o_number** attributes
+- Added support for new options from documentation **v2.015** (2025-02-19)
+  - Added **pickup_manipulation_lift** attributes
+  - Added **delivery_manipulation_lift** attributes
 
 
 ## [v7.12.0 (2024-11-20)](https://github.com/inspirum/balikobot-php/compare/v7.11.0...v7.12.0)